+++ examples/data-objects/flow-util-lib/src/index.ts
@@ -10,4 +10,3 @@ export { IView, View } from "./view";
 export { ICaretBounds, ICaretEvent, CaretEventType, Caret } from "./caret";
 export { Direction, getDeltaX, getDeltaY, getTabDirection, TabDirection } from "./direction";
 export { IRect, Rect } from "./rect";
-export { hasTagName, isElementNode, isTextNode, TagName } from "./tag";
