+++ examples/data-objects/webflow/src/editor/caret.ts
@@ -3,10 +3,10 @@
  * Licensed under the MIT License.
  */
 
-import { CaretEventType, Direction, getDeltaX, getDeltaY, ICaretEvent, hasTagName, TagName } from "@fluid-example/flow-util-lib";
+import { CaretEventType, Direction, getDeltaX, getDeltaY, ICaretEvent } from "@fluid-example/flow-util-lib";
 import { LocalReference } from "@fluidframework/merge-tree";
 import { DocSegmentKind, getDocSegmentKind } from "../document";
-import { clamp, Dom } from "../util";
+import { clamp, Dom, hasTagName, TagName } from "../util";
 import { ownsNode } from "../util/event";
 import { updateRef } from "../util/localref";
